Code review comments for remaining k8s reducer migration
