+++ frontend/packages/console-dynamic-plugin-sdk/src/api/common-types.ts
@@ -1,3 +1,5 @@
+import { Extension } from '../types';
+
 export { ResolvedExtension } from '../types';
 
 // Type for extension hook
@@ -199,3 +201,16 @@ export type DiscoveryResources = {
     };
   };
 };
+
+export type AlwaysOnExtension<P extends {} = any> = Omit<Extension<P>, 'flags'>;
+
+namespace ExtensionProperties {
+  export interface ModelDefinition {
+    /** Additional Kubernetes model definitions to register with Console. */
+    models: K8sModel[];
+  }
+}
+
+export interface ModelDefinition extends AlwaysOnExtension<ExtensionProperties.ModelDefinition> {
+  type: 'ModelDefinition';
+}
