+++ frontend/public/module/k8s/k8s-models.ts
@@ -4,12 +4,7 @@ import * as _ from 'lodash-es';
 // @ts-ignore
 import { useSelector } from 'react-redux';
 
-import {
-  K8sResourceKindReference,
-  K8sKind,
-  DiscoveryResources,
-  getModelExtensionMetadata,
-} from './index';
+import { K8sResourceKindReference, K8sKind, getModelExtensionMetadata } from './index';
 import * as staticModels from '../../models';
 import {
   referenceForModel,
@@ -21,6 +16,7 @@ import store from '../../redux';
 import { pluginStore } from '../../plugins';
 import { isModelDefinition, LoadedExtension } from '@console/plugin-sdk';
 import { isModelMetadata, ModelMetadata } from '@console/dynamic-plugin-sdk';
+import { DiscoveryResources } from '@console/dynamic-plugin-sdk/src/api/common-types';
 
 const modelKey = (model: K8sKind): string => {
   // TODO: Use `referenceForModel` even for known API objects
